refactor(gui): move tokensInUse out of the transport types file
`dashboard.ts` is a file of type declarations with one runtime function
living in it — a 23-line doc block over a 15-line function that reconstructs a
slot's token usage from whichever counters the llama-server build happens
to report.

The declarations around it are about to be replaced by generated ones.
The function cannot be, so it moves first: a mechanical swap should not
have to step around live logic, and a reviewer should not have to work out
which deletions are safe.

Pure relocation — the function, its doc, and the paragraph of the module
header that explains why this is an HTTP client of the proxy's own port
rather than a shared type. Its test moves with it and is renamed to match.
No behaviour changes and no logic is touched.
