+++ tests/ts/services/transport/slotTokens.test.ts
@@ -12,7 +12,8 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { tokensInUse, type SlotSnapshot } from '../../../../src/services/transport/types/dashboard';
+import { tokensInUse } from '../../../../src/services/transport/slotTokens';
+import type { SlotSnapshot } from '../../../../src/services/transport/types/dashboard';
 
 function slot(overrides: Partial<SlotSnapshot>): SlotSnapshot {
   return { id: 0, is_processing: false, ...overrides };
